rust: Rename module writable to writeable
