+++ src/bindings/rust/elektra/src/writeable.rs
